move from camel case to snake case
