+++ src/kibana/components/validate_query/lib/from_user.js
