+++ src/kibana/components/validate_query/lib/to_user.js
